pyproject.toml
[tool.poetry]
name = "cv-color-detector"
version = "0.1.0"
description = "A simple Computer Vision Color Detector"
authors = ["Owen Hochwald <oh@insomniacdesign.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
opencv-python = "^4.10.0.84"
numpy = "^2.2.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
